model: try fixing bone animations
Currently, animations are broken in such a way that flying birds
become streaks of grey. it appears to be an issue related to changed
interpolation: at some point noggit had a bug of only using 0% or 100%
of linear interpolation, but since that's fixed in
5efa814be65dbcd6676a37eb5c15a03193436439 they are now ugly. A fix
requires to find out what in interpolation or matrix multiplication is
broken.

This commit appears to fix things on the dwarf login screen. Birds in
the wild are also looking okay, but don't move for some reason. It
partially reverts 492ee69608f64604df5b636aa9b4ca3ba2e8b7e2 and tries
to be less smart.