ui: let tileset chooser be qt
Note: makes MapView constrution a 15 second issue. This is due to
rendering all tilesets in advance only. To get rid of that issue, one
should have a QStandardItem like class that has a lazy icon()
method. This needs an entire custom model and has thus been left as an
exercise to the reader.